removed 'pages' and got initial create post working
