--- backend/db/models/page.js
@@ -1,49 +1,0 @@
-'use strict';
-const {
-  Model,
-  Validator
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Page extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      this.belongsTo(models.User, { foreignKey: 'userId' });
-      // this.belongsTo(models.Theme, { foreignKey: 'defaultThemeId' });
-    }
-  }
-  Page.init({
-    userId: {
-      type: DataTypes.INTEGER,
-      references: {
-        model: 'Users',
-        key: 'id'
-      },
-      allowNull: false
-    },
-    defaultThemeId: {
-      type: DataTypes.INTEGER,
-      // references: {
-      //   model: 'Themes',
-      //   key: 'id'
-      // },
-    },
-    headerImg: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      defaultValue: '',
-    },
-  }, {
-    sequelize,
-    modelName: 'Page',
-    defaultScope: {
-      attributes: {
-        exclude: ["createdAt", "updatedAt"]
-      }
-    }
-  });
-  return Page;
-};
