+++ backend/db/models/post.js
@@ -12,8 +12,6 @@ module.exports = (sequelize, DataTypes) => {
      */
     static associate(models) {
       this.belongsTo(models.User, { foreignKey: 'userId' });
-      // this.belongsTo(models.Theme, { foreignKey: 'themeId' });
-      this.belongsTo(models.Page, { foreignKey: 'pageId' });
       this.hasOne(models.Song, { foreignKey: 'postId', onDelete: 'cascade', hooks: true });
       this.hasMany(models.Comment, { foreignKey: 'postId', onDelete: 'cascade', hooks: true });
       this.hasMany(models.Like, { foreignKey: 'postId', onDelete: 'cascade', hooks: true });
@@ -34,14 +32,6 @@ module.exports = (sequelize, DataTypes) => {
       //   model: 'Themes',
       //   key: 'id'
       // },
-      allowNull: false
-    },
-    pageId: {
-      type: DataTypes.INTEGER,
-      references: {
-        model: 'Pages',
-        key: 'id'
-      },
       allowNull: false
     },
     title: {
