+++ backend/db/models/theme.js
@@ -13,7 +13,6 @@ module.exports = (sequelize, DataTypes) => {
     static associate(models) {
       this.belongsTo(models.User, { foreignKey: 'userId' });
       // this.hasMany(models.Post, { foreignKey: 'themeId' });
-      // this.hasMany(models.Page, { foreignKey: 'DefaultThemeId', onDelete: 'cascade', hooks: true });
     }
   }
   Theme.init({
@@ -47,7 +46,7 @@ module.exports = (sequelize, DataTypes) => {
     borderstyle: {
       type: DataTypes.STRING,
       allowNull: false,
-      defaultValue: 'rounded'
+      defaultValue: 'solid'
     },
   }, {
     sequelize,
