--- backend/db/seeders/01-users.js:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\yourf\Downloads\express-skeleton.zip
