+++ backend/routes/api/likes.js
@@ -3,7 +3,7 @@ const bcrypt = require('bcryptjs');
 const { check } = require('express-validator');
 const { handleValidationErrors } = require('../../utils/validation');
 const { setTokenCookie, requireAuth } = require('../../utils/auth');
-const { User, Page, Comment, Like } = require('../../db/models');
+const { User, Comment, Like } = require('../../db/models');
 
 const router = express.Router();
