--- frontend/src/components/Pages/PageSetupForm.jsx
@@ -1,12 +1,0 @@
-import { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
-// import { postSongThunk } from "../../../redux/songs";
-import "./PageSetupForm.css";
-
-function PageSetupForm() {
-    const navigate = useNavigate();
-    const dispatch = useDispatch();
-    const user = useSelector((state) => state.session.user);
-
-}
