+++ frontend/src/redux/store.js
@@ -6,12 +6,10 @@ import {
 } from "redux";
 import thunk from "redux-thunk";
 import sessionReducer from "./session";
-import pagesReducer from "./pages";
 import postsReducer from "./posts";
 
 const rootReducer = combineReducers({
     session: sessionReducer,
-    pages: pagesReducer,
     posts: postsReducer,
 });
